pyproject.toml
[project]
name = "ai-agent-factory"
version = "0.1.0"
description = "Factory of AI agents for various tasks."
authors = [
    {name = "Challa Sharath", email = "csharath@outlook.com"},
]
dependencies = [
    "pyautogen==0.2.29",
    "python-dotenv>=1.0.1",
    "autogenstudio==0.1.2",
    "gpt-researcher==0.6.6",
]
requires-python = "==3.12.0"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false